📌 Restructured project dir, add readme credits
