+++ src/s21_matrix.h
@@ -20,6 +20,8 @@ typedef struct matrix_struct {
     int columns;
 } matrix_t;
 
+#define EPS 1e-6
+
 // TODO:
 int s21_calc_complements(matrix_t *A, matrix_t *result);
 int s21_determinant(matrix_t *A, double *result);
