+++ src/s21_sum_matrix.c
@@ -1,0 +1,12 @@
+#include "s21_matrix.h"
+
+int s21_sum_matrix(matrix_t *A, matrix_t *B, matrix_t *result) {
+    if (A->rows != B->rows || A->columns != B->columns)
+        return CALC_ERROR;
+
+    for (int i = 0; i < A->rows; i++)
+        for (int j = 0; j < A->columns; j++)
+            result->matrix[i][j] = A->matrix[i][j] + B->matrix[i][j];
+
+    return OK;
+}
